Switch from int to size_t to allow for a larger number of particles in an MPI domain.

+++ src/species_advance/species_advance.h
@@ -58,8 +58,8 @@ species_t *
 species( const char * name,
          float q,
          float m,
-         int max_local_np,
-         int max_local_nm,
+         size_t max_local_np,
+         size_t max_local_nm,
          int sort_interval,
          int sort_out_of_place,
          grid_t * g );
